Refactor some functions into a separate util module
